refactor(dream): split gates.py by concern to restore module-health headroom
gates.py sat exactly at the 500-LOC module-health cap (zero headroom — the next
non-comment line would trip "Split by concern"). Extract the impure acceptance-pass
orchestration + persistence (run_acceptance_pass, persist_probe_results,
_prior_pass_rate, _compliance_remediations) into a new sibling acceptance.py, leaving
gates.py as the pure gate computation (domain types, probe helpers, the Gate
predicates, evaluate_gates).

gates.py 500 -> 373 LOC; acceptance.py 154 LOC — both comfortably under the cap,
each well under the 10-public-function limit.

Pure refactor, no behavior change. The pure-gate public symbols stay importable
from gates unchanged. The two moved orchestration functions are imported from
acceptance by their consumers (phase_runner + test_gates/test_decay/test_phase_runner/
test_dream) rather than re-exported from gates: acceptance imports gates for the domain
types, so re-exporting the movers from gates would form a gates<->acceptance import
cycle.

Dependency direction is one-way: acceptance -> {gates, decay}. gates no longer
imports decay at runtime (its only decay use, cold_archive_names, moved with
run_acceptance_pass); decay's function-scoped imports of gate helpers (INDEX_BYTE_BUDGET,
_signature_line) are untouched, so that arrangement is preserved and no cycle is added.

Same 570-node dream test set passes; module-health, tach, intra-core (183) and
intra-loop (28) ratchets all green.
